requirements.txt
torch>=1.9.0
sentence-transformers>=2.2.0
pandas>=1.3.0
numpy>=1.19.5
qdrant-client>=1.1.0
pyyaml>=5.4.1
tqdm>=4.62.0
huggingface-hub>=0.19.0
tenacity>=8.0.1
openpyxl>=3.0.7